parse/requirements.txt
biopython
pandas
tenacity
lxml
